make signup page a component
